some RandMixin cleanup
cleaner to just put inside OpMixin
